japanese/py-pykakasi: New port: Lightweight converter from Japanese Kana-kanji sentences to Kana-Roman
pykakasi is a Python library that facilitates the conversion of Japanese text
from kanji to hiragana or katakana. It is designed for applications requiring
Japanese language processing, providing a straightforward interface for text
manipulation. This tool is lightweight and enhances accessibility for users
dealing with Japanese scripts.

WWW:	https://codeberg.org/miurahr/pykakasi

PR:	282369